# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Swaggerized Sails Boilerplate
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /thumbnail:
    x-swagger-router-controller: thumbnail
    post:
      summary: Patches a given data given a set of operations.
      operationId: generateThumbnail
      parameters:
        - $ref: '#/parameters/Header'
        - in: body
          required: true
          name: data
          description: The data to be sent.
          schema:
            type: object
            required:
              - imageURL
            properties:
              imageURL:
                type: string
                description: Public URL of the image
      responses:
        "200":
          description: Thumbnailing successful
          schema:
            $ref: "#/definitions/ThumbnailResponse"
        "400":
          description: Error creating thumbnail
          schema:
            $ref: "#/definitions/ThumbnailError"
        # responses may fall through to errors
        "default":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
parameters:
  Header:
    in: header
    name: Authorization
    type: string
    required: true
    description: Authorization Header to authenticate/authorize request
# complex objects have schema definitions
definitions:
...
  ThumbnailResponse:
    type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
      error:
        type: object
...
  ThumbnailError:
    required:
      - message
      - error
    properties:
      message:
        type: string
      error:
        type: object